requirements.txt
bs4
tmdbv3api
dateparser
python-dotenv
fastapi
uvicorn
tzdata
lxml
slowapi
curl_cffi
fake-headers
pillow
opencv-python-headless
pytesseract